feature(diverse): add cardlayout and map editor
I implemented a CardLayout to switch between the menu, settings and
play panels. The switch is temporarily done by hitting different arrow
keys. I also built the models and views for implementing a map
level builder. Currently changing the values in themap  array
in Map changes the tiles displayed in the PlayPanel. 0 is grass, 1 is
flower, 2 is path.